backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.30.6

# ── HTTP client (async) ────────────────────────────────────────────────────────
httpx==0.27.2

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# ── Settings & validation ──────────────────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1

# ── Security ───────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Utilities ──────────────────────────────────────────────────────────────────
python-multipart==0.0.20    # Form data / file uploads
python-dateutil==2.9.0       # Flexible date parsing
pytz==2024.2                 # Timezone support

# ── Development / testing ──────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
httpx                        # also used by TestClient
